requirements.txt
javalang==0.13.0
parsy==1.4.0
python-dotenv==0.21.0
lark==1.1.4 
pytest==7.2.0
psutil